+++ packages/stylelint-config/scss.js
@@ -18,14 +18,6 @@ module.exports = {
 			},
 		],
 
-		'block-opening-brace-space-before': 'always',
-		'block-closing-brace-newline-after': [
-			'always',
-			{
-				ignoreAtRules: [ 'if', 'else' ],
-			},
-		],
-		'at-rule-name-space-after': 'always',
 		'scss/at-else-closing-brace-newline-after': 'always-last-in-chain',
 		'scss/at-else-closing-brace-space-after': 'always-intermediate',
 		'scss/at-else-empty-line-before': 'never',